Day1_Classical_Physics/resources/requirements_presentation.txt
fpdf2>=2.7.5
matplotlib>=3.4.0
numpy>=1.21.0
